Split task_handler_helpers.go into focused files
Extract invocation interface and implementations into invocation.go,
base handler and options into task_handler_base.go, and move
generateCallbackToken to operation_tasks.go.
